+++ app/src/test/java/br/com/bg7/appvistoria/sync/FailWhenOnFailureCalledCallback.java
@@ -17,7 +17,7 @@ class FailWhenOnFailureCalledCallback implements SyncCallback {
     }
 
     @Override
-    public void onProgressUpdated(Inspection inspection, double progress) {
+    public void onProgressUpdated(Inspection inspection, Integer progress) {
 
     }
